Restructure datasources
