+++ packages/client/src/components/datasources/webpage/WebpageDatasource.vue
@@ -1,30 +1,30 @@
 <template>
   <div id="container">
     <div class="grid-data-card">
-      <div v-on:click="$emit('expandText')" id="grid-data-card-div-1">
-        <icon-file icon-color="#2ecc71"/>
+      <div v-on:click="$emit('expandWebpage')" id="grid-data-card-div-1">
+        <icon-web/>
       </div>
-      <div id="header" v-on:click="$emit('expandText')">Text Files</div>
+      <div id="header" v-on:click="$emit('expandWebpage')">Webpages</div>
       <div>
         <icon-min @click="add=!add" class="add" v-if="add"/>
         <icon-add @click="add=!add" class="add" v-else />
       </div>
     </div>
     <div v-if="add">
-      <AddDataURI placeholder-path="Enter Text File URI..." endpoint="http://localhost:3001/textdatasources"></AddDataURI>
+      <AddDataURI placeholder-path="Enter Webpage Link..." endpoint="http://localhost:3001/webpagedatasources"></AddDataURI>
     </div>
   </div>
 </template>
 
 <script>
-import AddDataURI from "./AddDataURI";
-import IconFile from "../icons/IconFile";
+import AddDataURI from "../AddDataURI";
 import axios from "axios";
-import IconMin from "../icons/IconMin";
-import IconAdd from "../icons/IconAdd";
+import IconWeb from "../../icons/IconWeb";
+import IconMin from "../../icons/IconMin";
+import IconAdd from "../../icons/IconAdd";
 
 export default {
-  name: "FileDataSource",
+  name: "WebpageDatasource",
   data() {
     return {
       expanded: false,
@@ -33,13 +33,13 @@ export default {
     }
   },
   components: {
-    IconAdd,
     IconMin,
-    IconFile,
+    IconAdd,
+    IconWeb,
     AddDataURI
   },
   beforeMount() {
-    axios.get("http://localhost:3001/textdatasources").then(
+    axios.get("http://localhost:3001/webpagedatasources").then(
         resp => {
           console.log(resp.data)
           this.dataSources = resp.data
@@ -86,6 +86,7 @@ export default {
 #grid-data-card-div-1 {
   margin-bottom: 8px
 }
+
 
 
 </style>
